+++ packages/drawer/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.3.1](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-drawer@7.3.0...@pluralsight/ps-design-system-drawer@7.3.1) (2020-10-13)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-drawer
+
+
+
+
+
+# [7.3.0](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-drawer@7.2.16...@pluralsight/ps-design-system-drawer@7.3.0) (2020-10-13)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-drawer
+
+
+
+
+
+## [7.2.16](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-drawer@7.2.15...@pluralsight/ps-design-system-drawer@7.2.16) (2020-10-13)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-drawer
+
+
+
+
+
 ## [7.2.16](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-drawer@7.2.14...@pluralsight/ps-design-system-drawer@7.2.16) (2020-10-09)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-drawer